demo/pyproject.toml
[tool.poetry]
name = "mimpaku"
version = "0.1.0"
description = "convert textfile to audio data script"
authors = ["TsuMakoto <tsukamoto@fusic.co.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.10"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
